requirements.txt
alembic==1.12.1
amqp==5.2.0
Babel==2.13.1
bcrypt==4.0.1
billiard==4.2.0
bleach==6.1.0
blinker==1.7.0
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
flask-babel==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
flask-mailman==1.0.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
Mako==1.3.0
MarkupSafe==2.1.3
mkdocs-material-extensions==1.3
mysql-connector-python==8.2.0
packaging==23.2
passlib==1.7.4
pdfkit==1.0.0
prompt-toolkit==3.0.40
protobuf==4.21.12
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.9
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
